LayoutTests/accessibility/ios-simulator/force-user-scalable.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
</script>
<meta name="viewport" content="width=device-width, user-scalable=no, maximum-scale=3, minimum-scale=1">
</head>
<body id="body">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that when we set force always user scalable, the maximum scale will be changed too.");

    if (window.accessibilityController) {
        window.internals.setViewportForceAlwaysUserScalable(true);
        shouldBe("window.internals.viewportConfigurationMaximumScale()", "10");
    }
    
    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
