LayoutTests/accessibility/ios-simulator/force-user-scalable-expected.txt

This tests that when we set force always user scalable, the maximum scale will be changed too.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.viewportConfigurationMaximumScale() is 10
PASS successfullyParsed is true

TEST COMPLETE
